NCBC-1330: Add CLR and OS to hello key
MOTIVATION
----------
When the client first makes contact with a cluster is sends a Hello
operation that includes some information about the client. Currently
only the client version is included but it would be useful to also
include the .NET CLR and the OS name.

MODIFICATIONS
-------------
- Update the Hello operation to add the CLR version and OS name to the
  hello operation key
- Add unit test to ensure the correct values are added for both .NET
  framework and .NET Core apps

RESULT
------
When sending a hello operation to the cluster the key contains the
couchbase client version, .NET CLR version and OS name & version the
client is running on.

Change-Id: Iaf2ba19daa90a809efeb170f1e455ff3945e2db3
Reviewed-on: http://review.couchbase.org/74019
Tested-by: Mike Goldsmith <goldsmith.mike@gmail.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>