NCBC-1334: Push client zip file to S3 after repository tag
MOTIVATION
----------
When AppVeyor builds after each commit, a zip file is created that
contains the client assemblies. These zip files are manually uploaded to
S3 but AppVeyor can do this automatically for a repository tag.

MODIFICATIONS
-------------
- add a new deployment action to appveyor.yml for S3

RESULT
------
After AppVeyor has built the zip file for a repository tag build it will
push the zip to S3.

Change-Id: I8afacb22f6f4a1bbfa6608277f786b0de3cf4887
Reviewed-on: http://review.couchbase.org/74031
Tested-by: Mike Goldsmith <goldsmith.mike@gmail.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>