NCBC-1346: Remove unsupported LookupInBuilder.Exists test
MOTIVATION
----------
As part of the DataStructures API integration tests were added to test
different scenarios. One of the tests is to verify in a document with
an array of strings, LoookupInBuilder.Exists can find a given value.
This is not a supported operation where LookupInBuilder.Exists can only
verify of a object property exits.

The failing test should be removed.

MODIFICATIONS
-------------
- Remove the LookupIn_Exists integration test

RESULT
------
Integration test for an unsupported operation are no longer executed
and fail.

Change-Id: Icb44f4475fd8e0c7640ab980441dd1562d92d64f
Reviewed-on: http://review.couchbase.org/74602
Tested-by: Mike Goldsmith <goldsmith.mike@gmail.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>