NCBC-1366: Fast-failover support
Motivation
----------
The upcoming Couchbase Server 5.0 offers a new feature called 'Fast
Failover' which allows failover to be configurable as low as 5s on the
server. This patch includes client-side support for FF by adding logic to
ensure that a config check only happens every 10ms to reduce spamming.

**NOTE: a future patch will be released as a .minor which will reduce the
HeartbeatConfigInterval from 10s to 2.5s to support the lower range of
support for FF on the server.

Modifications
-------------
 - Add ConfigMonitor class that encapsulates the logic for checking for
   configs at set intervals using CCCP only.
 - Add a centralized method on ClusterController called CheckConfigUpdate
   that can be called in the IO layer when certain classes of errors are
   encountered (Socket errors, etc).
 - Add a field called LastConfigCheckedTime which is used to ensure that
   there is a floor between checks so that the server is not spammed.
 - Add Endpoint to SyncState and bubble up so that the failed node can be
   omitted from the config check

Result
------
The SDK will check for a config update every 10s or when an IO
related error is detected. To reduce load on the server, the config check
will be limited to a single thread and will honor a floor of 10ms between
checks.

Change-Id: I7bc2ffa75bccf172f80d61b44130a82b1e8edc6f
Reviewed-on: http://review.couchbase.org/74621
Reviewed-by: Mike Goldsmith <goldsmith.mike@gmail.com>
Tested-by: Build Bot <build@couchbase.com>