NCBC-1348: Fix single sub-doc operations against small values
Motivation
----------
If performing a Counter sub-document mutation against a samll value,
without any other mutations in the same operation, and with enhanced
durability disabled, the result throws an NRE when you try to read the
new counter value.

Similar results when trying to read short values with sub-document Get
operations, but for these it fails regardless of the enhanced durability
setting.

Modifications
-------------
Override GetExtras in SubDocSingularLookupBase to do nothing, since
lookup sub-document operations don't return extras after the 24-byte
header.

Override GetExtras in SubDocSingleMutationBase to use the MultiMutation
implementation of GetExtras which only looks for the MutationToken when
VBucket is non-null.

Results
-------
For all singular sub-doc operations extras are now read correctly,
allowing the value to be read correctly afterwards.  This results in the
Bytes property being populated so that calls to Content don't throw an
NRE when trying to deserialize the value.

Change-Id: Ic90b1cacdace5612f50219939dd7bf60acc81c51
Reviewed-on: http://review.couchbase.org/74641
Reviewed-by: Mike Goldsmith <goldsmith.mike@gmail.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>