NCBC-1350: Fix query NRE with DefaultValueHandling.PopulateAndIgnore
Motivation
----------
If you use a custom deserialization setting of
DefaultValueHandling.PopulateAndIgnore, and N1QL query response which
doesn't include a "results" attribute (such as bad syntax) will throw an
NRE.

Modifications
-------------
Check for null "results" attribute and convert to an empty list when
creating the QueryResult<T> object.

Add an integration test for this issue.

Results
-------
NRE is prevented when DefaultValueHandling.IgnoreAndPopulate is used in
DeserializerSettings.

Change-Id: Ifb2e800c0f92d55d4eee95630356a2ce8da1f65c
Reviewed-on: http://review.couchbase.org/74657
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>