NCBC-1347: Convert N1QL query and prepare can cancel integration tests to unit tests
MOTIVATION
----------
The N1QL prepare and query methods both can take a custom
CancellationToken to stop a request before it has completed. Integration
tests were added to verify the behaviour but these intermittently fail
because of very short / non-deterministic query times. These can be made
into unit tests so they are more predictable.

MODIFICATIONS
-------------
- Move QueryAsync_CanCanel and PrepareAsync_CanCanel integation tests to
  be unit tests
- Check for and throw Exception if task was cancelled in
  FakeHttpMessageHandler

RESULT
------
The tests to verify if either a direct n1ql query or prepare query is
cancelled the result indicates failure and the exception is of type
OperationCancelledException. They are unit tests which do not depend on
an external Couchbase cluster.

Change-Id: Ibbe5d7cb606f647437a10311949ff97aa1b97f0a
Reviewed-on: http://review.couchbase.org/74982
Tested-by: Mike Goldsmith <goldsmith.mike@gmail.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>