NCBC-1353: Deploy nuget package on repository tag
MOTIVATION
----------
We create a nuget package each time AppVeyor is triggered and is already
pushed to MyGet. It would be good if we pushed the package to NuGet.org
on repository tag.

MODIFICATIONS
-------------
- Add new nuget deployment target to push nuget package on repo tag

RESULT
------
When the repository is tagged and when AppVeyor builds the project, the
nuget package is pushed to NuGet.

Change-Id: I389d3d1ecafe392aa85f245c434b84aebbd67ab2
Reviewed-on: http://review.couchbase.org/74925
Tested-by: Mike Goldsmith <goldsmith.mike@gmail.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>