NCBC-1356: Set user-agent header for all service requests (n1ql, fts, etc)
MOTIVATION
----------
When sending requests to HTTP service, such as N1QL and Search, it is
desirable to include the client description to help with debugging and
traceability.

MODIFICATIONS
-------------
- Add Utils/ClientIdentifier that returns the a description of the
  current client including assembly version, clr and os
- Use ClientIdentifer in Hello operations
- Set the User-Agent header when creating the CouchbaseHttpClient with
  the ClientIdentifier
- Add Unit test to ensure the User-Agent header value is set correct

RESULT
------
All requests sent to HTTP services set the User-Agent header with the
same client identifier that is used during Hello operations.

Change-Id: I4dcd4ffbbd7a8a1307eb6f3c5e7a24da62f3bbf4
Reviewed-on: http://review.couchbase.org/74978
Tested-by: Mike Goldsmith <goldsmith.mike@gmail.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>