NCBC-1357: Add IsConnected to FakeConnection in Couchbae.UnitTests
MOTIVATION
----------
The IConnection interface was recently updated to add a new IsConnected
property, however this was not implemented in the FakeConnection class
that implements the interface. This should be added with a default value
of true to allow the project to build.

MODIFICATIONS
-------------
- Add IsConnection that always returns true in FakeConnection

RESULT
------
The Couchbae.UnitTests project now builds successfully.

Change-Id: Ic6d964536366a7c732dc80754e48b3d810e42a85
Reviewed-on: http://review.couchbase.org/74979
Tested-by: Mike Goldsmith <goldsmith.mike@gmail.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>