NCBC-1351: Provide extensions to generate exceptions from responses
Motivation
----------
The approach of returning all errors as objects with success/failure
information is very powerful and flexible.  However, it can be somewhat
cumbersome for the consumer that simply wants an exception thrown if
there is an error.

Modifications
-------------
Add extension methods for each response type (both K/V types, view
query, N1QL query, and FTS query) which throw exceptions if the response
shows an error.  This exception includes details about the error
extracted from the response object.  If the response includs an
exception, this is made the inner exception of the new exception.

There is a specific exception type for each API (K/V, view, N1QL, and
FTS), all of which inherit from a base CouchbaseResponseException class.

Results
-------
Exceptions with error details can be easily thrown from responses if
that is desired in the consumer's implementation.  This loosely follows
the paradigm used by HttpClient's
HttpResponseMessage.EnsureSuccessStatusCode.

Example:
```
var response = bucket.GetDocument<dynamic>("key");
response.EnsureSuccess();
```

Change-Id: I4c26c0c48da3ea3b8578ee9a4a28a1e057d87f65
Reviewed-on: http://review.couchbase.org/74501
Reviewed-by: Mike Goldsmith <goldsmith.mike@gmail.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Build Bot <build@couchbase.com>