NCBC-1359: Fix AppVeyor script that removes InternalsVisibleTo entries
MOTIVATION
----------
When AppVeyor builds the Couchbase projects it removes 'friend' library
entries in the AssemblyInfo.cs as they are not needed outside of
testing. The script that does the removal is destroying the file which
leads to the assemblies failed to be corrected signed and have a default
version of 0.0.0.0.

MODIFICATIONS
-------------
- Fix the script that removes lines that includes "InternalVisibleTo" in
  AssemblyInfo.cs to use Set-Content instead of Out-File

RESULT
------
The AssemblyInfo.cs is updated as expected and the output assembly is
correctly signed and has the correct version.

Change-Id: I09c69ef7684134d2d697e91777d2ce5d30653ce9
Reviewed-on: http://review.couchbase.org/75005
Reviewed-by: Matt Ingenthron <matt@couchbase.com>
Tested-by: Mike Goldsmith <goldsmith.mike@gmail.com>