NCBC-1359: Wrap file reading in sub-process to prevent read/write race
MOTIVATION
----------
The script that reads and removes lines that contains
"InternalsVisibleTo" entries from AssemblyInfo.cs pipes it's commands
directly to one another. Occasionally it is possible to get a read/write
race where the file is still reading and when the last step is trying to
write which fails.

MODIFICATIONS
-------------
- Wrap the read section of the script into a sub-process by wrapping in
  parentheses

RESULT
------
The whole file is read into memory before the file and write actions to
ensure there is not a read/write race.

Change-Id: If9c1540239a46c8a834657705d4731feaec141a6
Reviewed-on: http://review.couchbase.org/75007
Reviewed-by: Matt Ingenthron <matt@couchbase.com>
Tested-by: Mike Goldsmith <goldsmith.mike@gmail.com>