NCBC-1341: Add RBAC authentication
MOTIVATION
----------
A new server feature is role based access control (RBAC) that provides
discrete user profiles with fine grained roles to services. The new
credential type utilises a discrete username and password for all
authentication processes.

Also, the original implementation does not follow the Cluster Level
Authentication RFC. It is likely it was changed post implementation and
should be updated to confirm to the updated RFC.

MODIFICATIONS
-------------
- Add IAuthenticator, RbacAuthenticator, ClassicAuthenticator and
  AuthenticatorType
- Add Cluster.Authenticate method that takes IAuthenticator
- Add SelectBucket operation and execute when retrieving bucket config
- Update ClusterController to separate bucket name and username fields
- Mark IClusterCredentials, ClusterCredentials and Cluster.Authenticate
  as Obsolete
- Add unit and integration tests to verify behaviour

RESULT
------
A new authenticator type is available that works with the new server
RBAC authentication process.

NOTE: The integration tests have been marked as ignored because the test
cluster is not capable of running those tests until has been updated.

Change-Id: Ib9dca2508e69f32c25a78ea30222bb7383ee8ed5
Reviewed-on: http://review.couchbase.org/74413
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>