NCBC-1355: Bump version to 2.4.3
MOTIVATION
----------
AppVeyor uses a build format to set the build version for non-tag
builds, this is useful for identifying what the next build number is in
the MyGet feed. This should be updated to 2.4.3 for the next SDK version
number.

MODIFICATIONS
-------------
- update build format to start with 2.4.3 in appveyor.yml

RESULT
------
AppVeyor builds for non-tag commits follow the build pattern
2.4.3.{build}.

Change-Id: I63a0acdc279069bbb63582092378091806e435d2
Reviewed-on: http://review.couchbase.org/75138
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>