NCBC-1380: Retrieve and load KV Error map
MOTIVATION
----------
Error Map is new server feature where a dictionary of error codes and
descriptions are supplied to the client from the server. This
information will contain additional errors that the client does not
understand directly.

This is the first stage implementation where the ErrorMap can be
enabled and retrieved from the server. Unknown response codes check the
error map for a match with the details being added to the result
message.

A future commit will add more functionality, such as retry behaviour.

MODIFICATIONS
-------------
- Add ResponseStatus.Failure to indicate a server error with additional
  details to come from the ErrorMap
- Add ResponseStatus.UnknownError to indicate an unknown status was
  received from the server
- Add UseKvError property to indicate if the IO service should try and
  get an ErrorMap
- Add SupportsErrorMap to indicate the server supports KV Error Maps
- Request the XError feature during Hello operations if UseErrorMap is
  true
- Retrieve Error map and store in IOService if hello response indicates
  server supports Error Map
- Add unit tests to show IO services can use the Error map for unknown
  response codes
- Add integration test to ensure the client requests and stores the
  error map if UseErrorMap is true

RESULT
------
If the client configuration enables KV Error map and the server supports
it, IO service connects to the server it will also request an Error Map.
This is stored in the IO service and passed into OperationBase.Read to
use if the response was non-successful.

The operation's message is updated with the additional information from
the error code.

Change-Id: Ie691ff8b4b69aa55f17c9a535130a63eeff002c7
Reviewed-on: http://review.couchbase.org/76321
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Build Bot <build@couchbase.com>