NCBC-1369: PoolConfiguration.MaxSize must come before MinSize
Motivation
----------
Fixes a bug where PoolConfigured.MaxSize must be assigned before
PoolConfiguration.MinSize or a ArgumentOutOfRangeException will be thrown.

Modifications
-------------
Add Validate method to PoolConfiguration that checks the values of MaxSize
and MinSize when the Initialize method of ClientConfiguration is called
internally.

Results
-------
The values are validated after assignment, so order is no longer
important.

Change-Id: I2ed654444aadb3ea1157cbd58125c06b5c1ac755
Reviewed-on: http://review.couchbase.org/75717
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Mike Goldsmith <goldsmith.mike@gmail.com>