NCBC-1370: MapGet throws NullReferenceException if key does not exist
Motivation
----------
If an error occurs, such as KeyNotFound, then a NullReferenceException
will be thrown when the body of the operation is deserialized.

Modifications
-------------
Check if the body is null and return the default value of the generic type
if it.

Result
------
If a KeyNotFound (or other error) occurs on a MapGet (or any other read
operation for Datastructures a NRE will not be thrown, instead the default
value for the generic parameter will be assigned to the Value field.

Change-Id: Ieee854f0f5d5b9ce78c3f80b66dd8c2bc7aee90d
Reviewed-on: http://review.couchbase.org/75727
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>
Reviewed-by: Mike Goldsmith <goldsmith.mike@gmail.com>