NCBC: 1352 Log bucket configuration JSON as parameter
MOTIVATION
----------
When CouchbaseRequestExecutor checks for updated configurations it
always logs the JSON. This seems to contain a string format
placeholder, eg {0}, and when the log goes through Log4Net to write
an error is thrown because there are no positional parameters that
match.

MODIFICATIONS
-------------
- pass the JSON string as a parameter

RESULT
------
When the log is being written, it should not throw an error.

Change-Id: I71d31268da837f8f17e4eaa741cc2a78b06280d4
Reviewed-on: http://review.couchbase.org/76392
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>