NCBC-1375: Default thread number to 3 when creating bucket connections
MOTIVATION
----------
When creating a bucket connection, a threadNumber property is sent in
the URI to indicate the number of threads to use for disk IO. This
currently defaults to 2, however the server takes this to indicate the
connection is to be treated as high-priority when it is should not.

The default should be 3 for low-priority, with 8 indicating
high-priority.

MODIFICATIONS
-------------
- remove ThreadNumber.Two enum entry
- replace all ThreadNumber.Two usage with ThreadNumber.Three

RESULT
------
When creating a bucket connection, the threadNumber defaults to 3, which
the server treats as a low-priority for disk IO.

Change-Id: Ic768f9b32ef1f70e241b34e4a917eafb6a62127f
Reviewed-on: http://review.couchbase.org/76490
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>