NCBC-1381: when UseConnectionPooling is false client still uses MuxIO
Motivation
----------
The ClientConfiguration.useConnectionPooling property when set to true,
should use connection pooling. This patch fixes a bug introduced in 2.4.0
that makes the client use MUXIO even when UseConnectionPooling is true.

Modification
------------
Added conditional check in ConnectionPoolingFactory to ensure that pooling
is used when the UseConnectionPooling property is true.

Results
-------
When UseConnectionPooling is true, the client will use connection pooling.

Change-Id: I36a52a70642c57a36ffb3a59204c6ecedb44d714
Reviewed-on: http://review.couchbase.org/76598
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Mike Goldsmith <goldsmith.mike@gmail.com>