NCBC-1379: Use appveyor variable for assembly patching
MOTIVATION
----------
For tagged builds the assembly versions should be updated
to use the tag name, however it being set to 0.0.0.0.

This is currently by patching the ProperyInfo.cs done using
the {version} variable. AppVeyor support suggests to use
an AppVeyor environment variable to set the assembly
version when patching.

MODIFICATIONS
-------------
- Use AppVeyor environment variable for assembly patching
- Disable publishing nuget publishing to nuget.org

RESULT
------
When AppVeyor builds, it correctly sets the assembly versions
for tagged builds.

Change-Id: Ibef9d8906326a9df174a9720dbbdf3fdeb52d969
Reviewed-on: http://review.couchbase.org/76621
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>