NCBC-1383: Connection fails to auth when using async K/V methods
Motivation
----------
Fixes a bug where the underlying socket connection may not be
authenticated causing an async operation to fail.

Modifications
-------------
Ensure that async Execute methods in the MultiplexingIOService check if
the socket is authenticated, if not then normal SASL auth will occur
before sending the K/V operation to the server.

Results
-------
If the connection was not authenticated by a previous sync call, it will
now be authenticatd by an async call.

Change-Id: I57b2bfe1fc35d871af553af77b88038af20084cb
Reviewed-on: http://review.couchbase.org/76672
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Mike Goldsmith <goldsmith.mike@gmail.com>