NCBC-1384: Ensure proper IIOService is created when UseConnectionPooling is true
Motivation
----------
Fixes a bug where a hybrid mux/pooled IIOService instance can be created
when using App.Config and UseConnectionPooling is true.

Modification
------------
Check if UseConnectionPooling is true, if it is force a regular connection
pool service to be created.

Result
------
If UseConnectionPooling is true, the client will use a pool service.

Change-Id: I80581072f20810d6a9da84ffd12e4a6f5a347de1
Reviewed-on: http://review.couchbase.org/76675
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Mike Goldsmith <goldsmith.mike@gmail.com>