NCBC-1387: Add advanced FTS searching options
MOTIVATION
----------
An update to the FTS RFC includes new advanced sorting options for
search queries. The new sorting options describe a SearchSort pattern
where 4 sorting options have been defined. The update also includes an
option to provide a raw JSON object for future custom sorting.

MODIFICATIONS
-------------
- Add ISearchSort, SearchSortBase, IdSearchSort, FieldSearchSort,
  ScoreSearchSort and GeoDistanceSearchSort
- Add two Sort overloads to ISearchParams that take either an
  ISearchSort or JObject
- Change _sort field to JArray in SearchParams
- Add sort implementations to SearchParams that store the sorting
  options in the _sort field
- Add unit tests to ensure both new Sort options return valid JSON
- Add unit tests to ensure all new SearchSort implementations return
  valid JSON and omit default values

RESULT
------
There are two new ways of applying sorting to search queries. Firstly by
using an implementation of ISearchSort and this commit includes 4
implementations. The second option is to provide a custom JObject which
describes the behaviour and is to support future custom sorting.

Change-Id: Ic51227d8aa3f5ec739027d186d26dc3854485960
Reviewed-on: http://review.couchbase.org/76953
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Mike Goldsmith <goldsmith.mike@gmail.com>