NCBC-1375: Re-add ThreadNumber.Two as Obsolete
MOTIVATION
----------
ThreadNumber.Two was removed in a previous commit, however this should
not be removed until the next major/minor release to ensure semver
compatibility.

MODIFICATIONS
-------------
- re-add ThreadNumber.Two enum entry with an obsolete attribute

RESULT
------
ThreadNumber.Two is still available for any applications that use it
directly but is marked as Obsolete to indicate it will be removed at the
next major/minor release.

Change-Id: Ib5dfb6004b13ceda1a1afac599aa7010376c7c07
Reviewed-on: http://review.couchbase.org/77109
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>