NCBC-1389: Property to test for ClusterHelper initialization
Motivation
----------
Provide a way to easily tell if ClusterHelper has been initialized.

Modifications
-------------
Added Initialized property which returns true when ClusterHelper is
initialized.

Ensured that _instance is always set to null on Close.

Added unit tests for the new property.

Results
-------
Initialized property now indicates if Initialize has been used, and
resets to false when Close is called.

Behavior of Close more consistent. The behavior was different if you had
Initialized but not used the ClusterHelper versus initializing and using
the ClusterHelper.  In the ununused state, ClusterHelper was left in a
state where it could be used again.  Now it is always unusable until the
next time you call Initialize.

Change-Id: Icd98f3223459dc581229ecdfc0c0f257b94ca7a1
Reviewed-on: http://review.couchbase.org/77228
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Mike Goldsmith <goldsmith.mike@gmail.com>