NCBC-1378: Make .NET client CouchbaseMock compatible
MOTIVATION
----------
CouchbaseMock (https://github.com/couchbase/couchbasemock) is a
testing tool that can simulate a Couchbase server functionality.
It would be useful for the .NET client to be compatible so we can
utilize it as a means of testing new functionality.

MODIFICATIONS
-------------
- Replace IPAddress with IPEndpoint in Config, VBucket and VBucket
  mappers
- Add default buckets config paths that can be used if the server
  does not respond with one
- Attempt verbose buckets config request if terse fails

RESULT
------
The .NET client is now compatible with CouchbaseMock and Integration
tests that target the default bucket can be run against a locally
running version of CouchbaseMock.

Change-Id: I840268c7c7eea43fb895fa7f0aab5dc0b6583a93
Reviewed-on: http://review.couchbase.org/77250
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Build Bot <build@couchbase.com>