NCBC-1394: OperationTimeout and high mem consumption when swapping node
Motivation
----------
While doing a failover/swap and rebalance, the client may get in a state
where its cluster configuration is not current resulting in every request
becoming a NMV. This leads to OperationTimeouts and high memory
consumption while the client tries to process the body of every NMV.

Modifications
-------------
Change from using a Monitor to a mutex so that every request is checked,
and no cluster map changes are skipped by the conditional monitor.

Result
------
Every cluster map revision will be checked, so there is no possibility of a
skipped server config. The client should recover from a failover/swap and
rebalance scenario.

Change-Id: I41303ca95e2d39fb658462024ce991f3b2150e2d
Reviewed-on: http://review.couchbase.org/77326
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jaekwon Park <jaekwon.park@couchbase.com>