NCBC-1398: Reference to ClientConfiguration is null in PoolConfiguration
Motivation
----------
The PoolConfiguration holds a reference to ClientConfiguration and in some
cases, depending upon how you build your configuration, this value can be
null. This will cause a NRE while bootstrapping.

Modifications
-------------
In the ClientConfiguration.Initialize() method we check to see if the
property is null, if it is then we set it to the current
ClientConfiguration reference.

Result
------
A NRE will not cause bootstrapping to fail if the ClientConfiguration is
null.

Change-Id: I3e22325257c5615d7c71a0f1a9cb58235e371c8f
Reviewed-on: http://review.couchbase.org/77358
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jaekwon Park <jaekwon.park@couchbase.com>