NCBC-1396: Only tear down once and load bucket before tests start
MOTIVATION
----------
The Memecached integration tests run very slowly on the CI server. This
is because the cluster is being disposed after each test run. Also,
memecached buckets are always slower to bootstrap because they only
support CCCP.

MODIFICATIONS
-------------
- Bootstrap the memcached bucket during setup
- Move the cluster dispose to after all tests have run

RESULT
------
The memcached tests run a faster because the cluster is not disposed
of after each test and the memcached bucket is loaded before any tests
are started.

Change-Id: I702d19795c3b1ff660f703bae75bd3b65c2f8c70
Reviewed-on: http://review.couchbase.org/77528
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>