NCBC-1410: Ensure bucketName is cloned for retries
Motivation
----------
If the BucketName is null for an operation when it fails and is retried, a
NullReferenceException will be caught and handled. This patch will ensure
that the bucektName is not null.

Modifications
-------------
Update all operations that are retriable so that the Clone() method clones
the bucketName as well.

Results
-------
Retriable operations will have a bucketName so that the response can be
mapped to the appropriate bucket if a config refresh is required.

Change-Id: I317c348af8edf3a88b46670dea6b6acbb9532538
Reviewed-on: http://review.couchbase.org/77629
Reviewed-by: Jaekwon Park <jaekwon.park@couchbase.com>
Tested-by: Jaekwon Park <jaekwon.park@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>