NCBC-1414: Add Cluster.Authenticate overload for username / password
MOTIVATION
----------
The Cluster.Authenticate method currently only takes an IAuthenticator
implementation. However, it would be convenient to add an overload that
takes the username and password as strings and creates a
PasswordAuthenticator internally.

MODIFICATIONS
-------------
- Add Cluster.Authenticate overload that takes username and password as
  strings
- Add unit tests to verify behaviour

RESULT
------
There is now an overload to the authenticate function on Cluster that
can take a username and password that internally creates a
PasswordAuthenticator for convenience.

Change-Id: I34f4443d0e5db67526e67ca37255be508eb278fe
Reviewed-on: http://review.couchbase.org/77710
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Mike Goldsmith <goldsmith.mike@gmail.com>