NCBC-1405: Add support for enhanced error messages
MOTIVATION
----------
As part of server 5.0, the server can optionally include additional
diagnostic information for a failed KV operation. The additional
information will be specific to that operation and may include details
to reference server side logs for security sensitive errors.

As an additional note, responses that do not include the JSON bit should
not be read and converted into a response message, which is the current
behaviour.

MODIFICATIONS
-------------
- Add DataType enum flags and set it when reading response header
- Use DataType enum instead of DataFormat on Header (previously unused)
- Extend OpeationBase.GetMessage to read and parse the response when it
  indicates the body is JSON
- Append context and reference data to operation message and exception
  data values
- Add unit tests to verify behaviour (missing values, default values,
  exceptions, etc)
- Add integration test that works with CouchbaseMock to verify behaviour

RESULT
------
Failed KV operations will now read the response body if the JSON bit is
set to try and find additional context and reference information. If
found, this information is appended to the response message and also to
any exception data if present.

NOTE: This commit does change behaviour where the response body was
always converted to a string and returned if there was no other
information available. This commit instead defaults to returning the
Status Code if no other information is available.

Change-Id: I730389a303043a7022a00d6b94ecb6c0f2448df9
Reviewed-on: http://review.couchbase.org/79816
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>