NCBC-1421: Ignore integration test for creating ephemeral buckets
MOTIVATION
----------
The test CreateBucket_DoesNotExist_Success integration test always fails
when run on the CI server because a 4.5 test cluster is used and does
not support Ephemeral buckets. This test should be ignored until the CI
server cluster is updated.

MODIFICATIONS
-------------
- Ignore CreateBucket_DoesNotExist_Success for
  TestCase(BucketTypeEnum.Ephemeral)

RESULT
------
The failing test does not get run automatically when the integration
tests are run on the CI server.

Change-Id: I2ca5e088a1ef5023bbb64d306c5fc6561e2d1bba
Reviewed-on: http://review.couchbase.org/77937
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>