NCBC-1419: Update readme with details on unit and integration testing
MOTIVATION
----------
The README.md is the first thing developers see when viewing the GitHub
repository and should be accurate. The README requires an update to
describe how the test projects are organized and how to run them.

MODIFICATIONS
-------------
- Update README.md to describe test projects and how to setup / run them

RESULT
------
The readme more accurately represents how the client's tests are
organized and how to correctly run them.

Change-Id: Idc129777090e448dfaf29b7576ba5dc568a5ae43
Reviewed-on: http://review.couchbase.org/78186
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>