NCBC-1429: NotSupportedException thrown when SSL enabled under high concurrency
Motivation
----------
Fixes a bug where a connection will be enqueued twice and then consumed on
two or more threads causing the SslStream to throw a NotSupportedException
because only a single thread can use the stream at one time.

Modifications
-------------
Added a mutex lock to the queue and do a check to see if the connection
has already been enqueued.

Results
-------
Subsequent runs of the unit test indicate no exceptions are thrown.

Change-Id: Iaf018325c4576853aa6314f83bead0ea01f68781
Reviewed-on: http://review.couchbase.org/78340
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Mike Goldsmith <goldsmith.mike@gmail.com>