NCBC-1432: Support Facets in FTS query responses
MOTIVATION
----------
Facets can be added to FTS queries for additional searching behaviour.
The submission of facets in a query has already been committed, however,
the parsing and processing of the results is missing. This commit adds
support for the three available facets: term, numeric range and date
range.

MODIFICATIONS
-------------
- Update ISearchResult.Facets to be a dictionary<string, IFacetResult>
- Update SearchDataMapper to parse and process facet results
- Add FacetResultType, IFacetResult, DefaultFacetResult,
  TermFacetResult, Term, NumericRangeFacetResult NumericRange,
  DateRangeFacetResut and DateRange
- Add unit and integration tests to verify behaviour

RESULT
------
Facet results are now supported and returned when included in an FTS
query.

Change-Id: Ie36150bfb51a515ff539e15d0f581962d29e6825
Reviewed-on: http://review.couchbase.org/78488
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>