NCBC-1434: Don't use IOService to authenticate PlainTextMechanism
MOTIVATION
----------
As part of a recent refactor, SASL authentication flow has changed with
authentication occurring before an IOService is created. This means that
SASL implementations can not depend on an IOService, and should instead
implement an internal way to authenticate a connection.

MODIFICATIONS
-------------
- Remove IOService references from PlainTextMechanism and implement
  Execute method
- Remove IOservice comment from CramMd5Mechanism
- Update PlainTextMechanismTests to not use IOService in constructor

RESULT
------
The PlainTextMechanism no longer depends on an IOService to
authenticate connections.

Change-Id: I434b71b177498b5c143fe014efdee54de0b402c8
Reviewed-on: http://review.couchbase.org/78919
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>