NCBC-1442: Refactor PooledIOService for SharedConnectionPool
Motivation
----------
SharedConnectionPool requires that the connection be explicitly released
back into the pool when the operation has completed IO so that the state
of connection can be checked to see if it is still good or perhaps needs
to be recreated if an IO or other error occurred. PooledIOConnection does
the same, but uses the older Connection type - for async operations the
connection itself will be released back into the pool when the call back
is called (its completed its IO) implicitly. The behavior difference
between explicitly or implicitly requires sub-classing or Type specific
logic; this commit provides specialization over Type specific logic.

Modifications
-------------
 - Added SharedPooledIOService class that derives from PooledIOService
 - Made async methods on PooledIOService virtual
 - Override async methods in SharedPooledIOService
 - Make shared fields protected
 - Make IOServiceFactory default to SharedConnectionPool

Results
-------
Both shared and non-shared pooled connections have the correct release
behavior.

Change-Id: I908d16c408a02750791a3495d08a2e97cde2a2a4
Reviewed-on: http://review.couchbase.org/79327
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Mike Goldsmith <goldsmith.mike@gmail.com>