NCBC-1445: Add RBAC config section integration tests
MOTIVATION
----------
It should be possible to load a client configuration using a
configuration section defined in an app or web config. This looks to
already be working, but should have tests to ensure the feature
continues to work.

MODIFICATIONS
-------------
- add tests for both PasswordAuthenticator and ClassicAuthenticator to
  ensure it can authenticate when loaded using client config section

RESULT
------
There is now test coverage to ensure loading a configuration from
app.config works as expected.

Change-Id: I1f2371f5440f9c3541c75e5e0d28e15fd19a0dee
Reviewed-on: http://review.couchbase.org/79715
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>