NCBC-1443: Provide SDK endpoints to request the cluster version
Motivation
----------
Provide a consistent means to determine the cluster version via the SDK.
This is particularly important since the introduction of RBAC in 5.0,
since the endpoints that provide the version now require authentication.

Modifications
-------------
Created ImplementationVersion structure to represent the parsed
implementation version, with support for comparison operators.

Added ConfigInfo property to IConfigObserver to provide a way to access
the current configuration for a bucket.

Created internal ClusterVersionProvider class to handle getting and
parsing the implementation version at either the bucket or cluster
level.

Added GetImplementationVersion sync and async methods to IBucket and
ICluster.

Created unit and integration tests.

Results
-------
The cluster's implementation version can now be requested via an
ICluster or IBucket.  It is important to support both for the sake of
backwards compatibility.  Consumers implemented prior to 5.0 will only
have bucket-level authentication configured, using the bucket name as
the user name.  If they are targeting a 5.0 server, they'll need to
request the version via IBucket so it can use the bucket credentials.

Change-Id: I6c52821fdfadb25c089ffbbf3b3399cdc4cee499
Reviewed-on: http://review.couchbase.org/79350
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Mike Goldsmith <goldsmith.mike@gmail.com>